% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm.bootstrap.functions.R
\name{gbm.bootstrap.functions}
\alias{gbm.bootstrap.functions}
\title{gbm.bootstrap.functions}
\usage{
gbm.bootstrap.functions(gbm.object, list.predictors, n.reps = 200,
  n.divisions = 100, verbose = T)
}
\arguments{
\item{gbm.object}{a gbm object describing sample intensity}

\item{list.predictors}{a list of data from a run of plot.gbm.4list}

\item{n.reps}{number of bootstrap samples}

\item{n.divisions}{the number of used in continuous resolution of plot.gbm}

\item{verbose}{control reporting}
}
\description{
Function to calculate bootstrap estimates of 95% confidence intervals for partial plots of a brt model fitted with gbm.step. See e.g. Buston and Elith 2011 (J. Anim. Ecol). 'gbm.bootstrap.functions', 'plot.gbm.4list' and 'plot.gbm.boot' were originally written by Jane Elith and John Leathwick, but not released publicly. We thank them for providing this code.
}
