% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCatInfluence.R
\name{ggCatInfluence}
\alias{ggCatInfluence}
\title{ggCatInfluence}
\usage{
ggCatInfluence(..., category = NULL, data = df, signif = FALSE)
}
\arguments{
\item{...}{several gbm.step object (object of S3 class gbm)}

\item{category}{the name of the column that contains the different categories associated to the predictors}

\item{signif}{Logical. If TRUE (default FALSE), retains only the "significant" predictors (Muller et al. 2013) and rescale their influence to 100}

\item{df}{a dataframe that contains the predictors and their categories}
}
\description{
Function to return the relative influence per category of predictors from boosted regression trees obtained with the gbm.step routine in the dismo package.
}
