% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInfluence.R
\name{ggInfluence}
\alias{ggInfluence}
\title{ggInfluence}
\usage{
ggInfluence(gbm.object, col.bar = "skyblue3", show.signif = TRUE,
  col.signif = "#8B3A3A", main = gbm.call$response.name,
  x.label = "Relative influence (\%)", signif = FALSE, plot = T, ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{col.bar}{color of the bars}

\item{show.signif}{Logical. If TRUE (default), display a vertical dashed line indicating the threshold of significance (100/number of predictors, Muller et al. 2013)}

\item{col.signif}{color of the dashed line}

\item{main}{title for the plot (by default the name of the response variable)}

\item{x.label}{title for the x-axis}

\item{signif}{Logical. If TRUE (default FALSE), retains only the "significant" predictors (Muller et al. 2013) and rescale their influence to 100}

\item{plot}{Logical. If TRUE (default) draws a horizontal barplot of the relative influence of each predictor}
}
\value{
Returns a dataframe with the relative influence of each predictor in the model and a barplot
}
\description{
Function to return and plot the relative influence of predictors from boosted regression trees obtained with the gbm.step routine in the dismo package.
}
