% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInteract_2D.R
\name{ggInteract_2D}
\alias{ggInteract_2D}
\title{ggInteract_2D}
\usage{
ggInteract_2D(gbm.object, dat.rank = NULL, n.plots = NULL, x = 1, y = 2,
  col.gradient = c("lightskyblue", "#142c42"), palette = NULL,
  show.axis = F, show.dot = F, contour = T, label.contour = F,
  col.contour = "cadetblue2", col.dot = "grey10", alpha.dot = 0.9,
  cex.dot = 0.5, pred.means = NULL, x.label = NULL, y.label = NULL,
  legend = TRUE, z.label = "Fitted value", x.range = NULL, nrow = NULL,
  ncol = NULL, y.range = NULL, z.range = NULL, main = "",
  smooth = "none", ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{dat.rank}{allows to plot several interaction plots using a data frame of interaction size (e.g. obtained from ggInteract_list)}

\item{n.plots}{number of interaction plots to display, ranked by decreasing interaction size}

\item{x}{in case of a single plot, the first predictor. Can be either a character name (use "") or a number indicating its index}

\item{y}{in case of a single plot, the second predictor. Can be either a character name (use "") or a number indicating its index}

\item{col.gradient}{a vector of two colors to define the color gradient of the plot}

\item{palette}{use a predefined color palette (default = NULL)}

\item{show.dot}{Logical. If TRUE (default=FALSE), overlays the actual data points}

\item{contour}{Logical. If TRUE (default), draw contours}

\item{label.contour}{Logical. If TRUE (default=FALSE), adds labels to the contours}

\item{col.contour}{color of the contour lines}

\item{col.dot}{color of the dots}

\item{alpha.dot}{transparency of the dots}

\item{cex.dot}{size of the dots}

\item{pred.means}{allows specification of values for other variables (see gbm.perspec in dismo)}

\item{x.label}{title for the x-axis}

\item{y.label}{title for the y-axis}

\item{legend}{Logical. If TRUE (default) shows the legend}

\item{z.label}{title for the z-axis}

\item{x.range}{manual range for the x-axis}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per column}

\item{y.range}{manual range for the y-axis}

\item{z.range}{manual range for the z-axis}

\item{main}{title for the plot}

\item{smooth:}{controls smoothing of the predicted surface}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per column}
}
\description{
Function to draw 2D interaction plot showing predicted values for two predictors from a boosted regression trees obtained with the gbm.step routine in the dismo package. Values for all other variables are set at their mean by default.
}
