% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInteract_3D.R
\name{ggInteract_3D}
\alias{ggInteract_3D}
\title{ggInteract_3D}
\usage{
ggInteract_3D(gbm.object, x = 1, y = 2, col.gradient = "Blues",
  legend = T, main = "", x.label = NULL, y.label = NULL,
  x.range = NULL, y.range = NULL, z.label = "Fitted value",
  z.range = NULL, pred.means = NULL, smooth = "none", ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{x}{the first predictor. Can be either a character name (use "") or a number indicating its index}

\item{y}{the second predictor. Can be either a character name (use "") or a number indicating its index}

\item{col.gradient}{a vector of two colors to define the color gradient of the plot or a predefined palette}

\item{legend}{Logical. If TRUE (default) shows the legend}

\item{main}{title for the plot}

\item{x.label}{title for the x-axis}

\item{y.label}{title for the y-axis}

\item{x.range}{manual range for the x-axis}

\item{y.range}{manual range for the y-axis}

\item{z.label}{title for the z-axis}

\item{z.range}{manual range for the z-axis}

\item{pred.means}{allows specification of values for other variables (see gbm.perspec in dismo)}

\item{smooth}{controls smoothing of the predicted surface (default= "none")}
}
\description{
Function to draw 3D interaction plot showing predicted values for two predictors from a boosted regression trees obtained with the gbm.step routine in the dismo package. Values for all other variables are set at their mean by default. Possibility to export an interactive version of the plot in html format.
}
