% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInteract_boot_hist.R
\name{ggInteract_boot_hist}
\alias{ggInteract_boot_hist}
\title{ggInteract_boot_hist}
\usage{
ggInteract_boot_hist(data = NULL, column = NULL, cis = c(0.025, 0.975),
  bindwidth = 1, obs = NULL)
}
\arguments{
\item{data}{a dataframe that contains the values}

\item{column}{which column to plot}

\item{cis}{a vector to specify the confidence intervals (default=c(0.025, 0.975))}

\item{bindwidth}{bindwidth of the barplot (default = 1)}

\item{obs}{the observed value to compare to the distribution and confidence intervals}
}
\description{
Function to plot a distribution histogram with confidence intervals
}
