% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInteract_list.R
\name{ggInteract_list}
\alias{ggInteract_list}
\title{ggInteract_list}
\usage{
ggInteract_list(gbm.object, index = T, global.env = F)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{index}{Logical. If TRUE (default) returns the interactions table with the index of each predictors in addition to their names}

\item{global.env}{Logical. If TRUE (default FALSE) exports the result into the R global environment as an object called  Interact_list. Be aware of the critics of exporting to the global environment}
}
\value{
Returns a dataframe with the interaction sizes ranked from largest to smallest
}
\description{
Function to return the interactions size for a given boosted regression trees obtained with the gbm.step routine in the dismo package.
}
