% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMulti.list.R
\name{ggMulti.list}
\alias{ggMulti.list}
\title{ggMulti.list}
\usage{
ggMulti.list(gbm.object, variable.no = 0, n.plots = length(pred.names),
  y.label = "Fitted function", x.label = NULL, ...)
}
\description{
Function used in ggMultiPD and ggMultiPDfit to display results from different BRTs (with same predictors) on a single plot
}
