% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMultiInfluence.R
\name{ggMultiInfluence}
\alias{ggMultiInfluence}
\title{ggMultiInfluence}
\usage{
ggMultiInfluence(..., col.gradient = c("white", "lightblue4"), round = 1,
  col.text = "grey10", col.grid = NULL, size.grid = 0.3,
  legend.pos = "bottom", legend.dir = "horizontal", scale.gradient = c(0,
  100))
}
\arguments{
\item{...}{several gbm.step objects (object of S3 class gbm)}

\item{col.gradient}{a vector of two colors to define the color gradient of the plot}

\item{round}{indicates the number of decimal places to be used to round the relative influence values (default = 1)}

\item{col.text}{color of the text within the grid cells}

\item{col.grid}{color of the contours of the cells (default= NULL)}

\item{size.grid}{thickness of the contours of the cells (if col.grid is provided)}

\item{legend.pos}{position for the legend ("none", "right", "left", "bottom" (default), "top")}

\item{legend.dir:}{direction for the legend ("vertical", "horizontal" (default))}

\item{scale.gradient:}{a vector with min and max limits for the gradient values (default=c(0,100))}
}
\value{
Returns a dataframe with the relative influence of each predictor for the different models and plot a heatmap
}
\description{
Function to return and plot (heatmap) the relative influence of predictors from several boosted regression trees obtained with the gbm.step routine in the dismo package. The models need to have the same set of predictors.
}
