% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMultiPD.R
\name{ggMultiPD}
\alias{ggMultiPD}
\title{ggMultiPD}
\usage{
ggMultiPD(..., n.plots = length(pred.names), predictor = NULL,
  col.lines = c(), cex.line = 0.5, legend.pos = "right", smooth = F,
  cex.smooth = 0.3, se = F, span = 0.3, nrow = NULL, ncol = NULL)
}
\arguments{
\item{...}{several gbm.step object (object of S3 class gbm)}

\item{n.plots}{number of predictors to display, ranked by decreasing relative influence (default=all)}

\item{predictor}{select a single predictor to plot. Can be either a character name (use "") or a number indicating its ranking in terms of relative influence}

\item{col.lines}{a vector with as many colors as there are BRTs}

\item{cex.line}{size of the lines}

\item{legend.pos}{position for the legend ("none", "right", "left", "bottom", "top")}

\item{smooth}{Logical. If TRUE (default FALSE) add a smoother to each fitted function}

\item{cex.smooth}{size of the smoothers}

\item{span}{span of the smoothers (default = 0.3)}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per column}
}
\description{
Function to draw partial dependency plots with fitted functions from several boosted regression trees obtained with the gbm.step routine in the dismo package. The models need to have the same set of predictors.
}
