% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMultiPDfit.R
\name{ggMultiPDfit}
\alias{ggMultiPDfit}
\title{ggMultiPDfit}
\usage{
ggMultiPDfit(..., n.plots = length(pred.names), predictor = NULL,
  col.dots = c(), alpha.dot = 0.5, cex.dot = 1, legend.pos = "right",
  smooth = F, cex.smooth = 0.3, se = F, span = 0.3, nrow = NULL,
  ncol = NULL)
}
\arguments{
\item{...}{several gbm.step object (object of S3 class gbm)}

\item{n.plots}{number of predictors to display, ranked by decreasing relative influence (default=all)}

\item{predictor}{select a single predictor to plot. Can be either a character name (use "") or a number indicating its ranking in terms of relative influence}

\item{col.dots}{a vector with as many colors as there are BRTs}

\item{alpha.dot}{transparency of the dots (default = 0.5)}

\item{cex.dot}{size of the dots}

\item{legend.pos}{position for the legend ("none", "right", "left", "bottom", "top")}

\item{smooth}{Logical. If TRUE (default FALSE) add a smoother to each set of fitted values}

\item{cex.smooth}{size of the smoothers}

\item{se}{Logical. If TRUE (default FALSE), add standard error around the smoother}

\item{span}{span of the smoothers (default = 0.3)}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per column}
}
\description{
Function to draw partial dependency plots with fitted values from several boosted regression trees obtained with the gbm.step routine in the dismo package. The models need to have the same set of predictors.
}
