% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPD.R
\name{ggPD}
\alias{ggPD}
\title{ggPD}
\usage{
ggPD(gbm.object, n.plots = length(pred.names), predictor = NULL,
  nrow = NULL, ncol = NULL, col.line = "darkorange", cex.line = 0.5,
  smooth = FALSE, col.smooth = "blue", cex.smooth = 0.3, span = 0.3,
  rug = FALSE, rug.pos = "t", common.scale = TRUE,
  y.label = "Fitted function", x.label = NULL, ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{n.plots}{number of predictors to display, ranked by decreasing relative influence (default=all)}

\item{predictor}{select a single variable. Can be either a character name (use "") or a number indicating its ranking in terms of relative influence}

\item{ncol}{number of plots per column}

\item{col.line}{color of the line}

\item{cex.line}{size of the line (default = 0.5)}

\item{smooth}{Logical. If TRUE, (default FALSE) add a smoother to the fitted function}

\item{col.smooth}{color of the smoother}

\item{cex.smooth}{size of the smoother line (default = 0.3)}

\item{span}{span of the smoother (default = 0.3)}

\item{rug}{Logical. If TRUE (default FALSE), add rugs to the plot indicating the actual data points}

\item{rug.pos}{position of the rugs (default = "t"). Can be either "t" (top), "l" (left), "b" (bottom), or "r" (right)}

\item{common.scale}{Logical. If TRUE (default), all the plots have the same y limits}

\item{y.label}{title for the y-axis (default = "Fitted function")}

\item{x.label}{title for the x-axis (by default the predictor name and its relative influence)}

\item{nrow:}{number of plots per row}
}
\description{
Function to draw partial dependency plots with fitted function from a boosted regression trees obtained with the gbm.step routine in the dismo package.
}
