% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPDboot.R
\name{ggPD_boot}
\alias{ggPD_boot}
\title{ggPDboot}
\usage{
ggPD_boot(gbm.object, predictor = NULL, n.plots = length(pred.names),
  list.4.preds = NULL, booted.preds = NULL, nrow = NULL, ncol = NULL,
  col.line = "darkorange", cex.line = 0.5, type.ci = "lines",
  col.ci = "grey80", cex.ci = 0.3, lty.ci = 2, alpha.ci = 0.5,
  smooth = FALSE, col.smooth = "blue", cex.smooth = 0.3, span = 0.3,
  rug = FALSE, rug.pos = "t", common.scale = TRUE, cis = c(0.025,
  0.975), y.label = "Fitted function", x.label = NULL, ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{predictor}{select a single variable. Can be either a character name (use "") or a number indicating its ranking in terms of relative influence}

\item{n.plots}{number of predictors to display, ranked by decreasing relative influence (default=all)}

\item{list.4.preds}{the expand.grid list over variable range (from plot.gbm.4list)}

\item{booted.preds}{the array from the bootstrap run (from gbm.bootstrap.functions)}

\item{ncol}{number of plots per column}

\item{col.line}{color of the line}

\item{cex.line}{size of the line (default = 0.5)}

\item{type.ci}{display the CIs either as "lines" (default) or "ribbon"}

\item{col.ci}{color of the CIs}

\item{cex.ci}{size of the CIs (relevant for type.ci="lines")}

\item{lty.ci}{line types for the CIs (relevant for type.ci="lines")}

\item{col.smooth}{color of the smoother}

\item{cex.smooth}{size of the smoother line (default = 0.3)}

\item{span}{span of the smoother (default = 0.3)}

\item{rug}{Logical. If TRUE (default FALSE), add rugs to the plot indicating the actual data points}

\item{rug.pos}{position of the rugs (default = "t"). Can be either "t" (top), "l" (left), "b" (bottom), or "r" (right)}

\item{common.scale}{Logical. If TRUE (default), all the plots have the same y limits}

\item{cis}{a vector to specify the confidence intervals (default=c(0.025, 0.975))}

\item{y.label}{title for the y-axis (default = "Fitted function")}

\item{x.label}{title for the x-axis (by default the predictor name and its relative influence)}

\item{nrow:}{number of plots per row}

\item{alpha.ci:}{transparency of the ribbon (relevant for type.ci="ribbon")}

\item{smooth:}{Logical. If TRUE, (default FALSE) add a smoother to the fitted function}
}
\description{
Function to draw partial dependency plots with fitted function and confidence intervals obtained through bootstrapping (with replacement) from a boosted regression trees obtained with the gbm.step routine in the dismo package.
}
