% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPDfit.R
\name{ggPDfit}
\alias{ggPDfit}
\title{ggPDfit}
\usage{
ggPDfit(gbm.object, n.plots = length(pred.names), predictor = NULL,
  nrow = NULL, ncol = NULL, col.dot = "grey", cex.dot = 1,
  alpha.dot = 0.4, smooth = FALSE, col.smooth = "red", span = 0.3,
  se = F, cex.smooth = 0.3, y.label = "Fitted values", x.label = NULL,
  ...)
}
\arguments{
\item{gbm.object}{a gbm.step object (object of S3 class gbm)}

\item{n.plots}{number of predictors to display, ranked by decreasing relative influence (default=all)}

\item{predictor}{select a single variable. Can be either a character name (use "") or a number indicating its ranking in terms of relative influence}

\item{nrow}{number of plots per row}

\item{ncol}{number of plots per column}

\item{col.dot}{color of the dots}

\item{cex.dot}{size of the dots}

\item{alpha.dot}{transparency of the dots (default = 0.4)}

\item{smooth}{Logical. If TRUE, (default FALSE) add a smoother to the fitted values}

\item{col.smooth}{color of the smoother}

\item{span}{span of the smoother (default = 0.3)}

\item{se}{Logical. If TRUE (default FALSE), add standard error around the smoother}

\item{cex.smooth}{size of the smoother line (default = 0.3)}

\item{y.label}{title for the y-axis (default = "Fitted values")}

\item{x.label}{title for the x-axis (by default the predictor name and its relative influence)}
}
\description{
Function to draw partial dependency plots with fitted values from a boosted regression trees obtained with the gbm.step routine in the dismo package.
}
