% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPerformance.R
\name{ggPerformance}
\alias{ggPerformance}
\title{ggPerformance}
\usage{
ggPerformance(..., global.env = F)
}
\arguments{
\item{...}{one or several gbm.step objects (object of S3 class gbm)}

\item{global.env}{Logical. If TRUE (default FALSE) exports the result into the R global environment as an object called ModelPerf. Be aware of the critics of exporting to the global environment}
}
\value{
Returns a dataframe with mean total deviance, mean residual deviance, correlation, area under the receiver operating curve (AUC), percentage deviance explained, cross-validated deviance, cross-validated correlation, cross-validated AUC and cross-validated percentage deviance explained (calculated with the formula 1-(cross-valdiated deviance/Total.Deviance))
}
\description{
Function to extract the models performance of one or several boosted regression trees obtained with the gbm.step routine in the dismo package.
}
