% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gbm.4list.R
\name{plot.gbm.4list}
\alias{plot.gbm.4list}
\title{plot.gbm.4list}
\usage{
plot.gbm.4list(x, i.var = NULL, n.trees = x$n.trees,
  continuous.resolution = 100, return.grid = TRUE, ...)
}
\description{
'gbm.bootstrap.functions', 'plot.gbm.4list' and 'plot.gbm.boot' were originally written by Jane Elith and John Leathwick, but not released publicly. We thank them for providing this code.
}
